Add files via upload
Practicing using a clustering algorithm. K-Means is a popular clustering algorithm used in machine learning and data mining. It is an iterative algorithm that partitions a dataset into K distinct, non-overlapping clusters. The "K" in K-Means represents the number of clusters that the algorithm aims to create from the input data.